+++ src/test/java/com/alibaba/json/bvt/issue_2200/Issue2214.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.issue_2100;
+package com.alibaba.json.bvt.issue_2200;
 
 import com.alibaba.fastjson.JSON;
 import junit.framework.TestCase;
